requirements.txt
livekit-agents
livekit-plugins-google
livekit-plugins-noise-cancellation
livekit-plugins-silero
livekit
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
pyaudio